split game_row
